Add files via upload
In this problem, we have to find the shortest subsequence or subarray in an array that is sorted. Using a two pointer approach (first pointer (denoted by i) starting from i=0 and the second (denoted by j) from j=array.size()-1) we calculate the first and last indices of the unsorted array. The length of a subarray is calculated as j-i+1.
Time complexity is O(NLogN).